+++ src/ecoz2/ecoz2.c
@@ -7,7 +7,7 @@
 #include <stdlib.h>  // rand, srand
 
 const char *ecoz2_version() {
-    return "0.3.5";
+    return "0.3.6";
 }
 
 void ecoz2_set_random_seed(int seed) {