+++ src/hmm/distr.c
@@ -15,63 +15,61 @@ static inline prob_t randomValue() {
 }
 
 // Random distribution, with no null-values
-void dis_inicAle(prob_t *dis, int lon) {
-    prob_t acum = 0.;
-    for (int i = 0; i < lon; i++) {
-        acum += dis[i] = randomValue() + 1;
+void dis_set_random(prob_t *dis, int len) {
+    prob_t cumulative = 0.;
+    for (int i = 0; i < len; i++) {
+        cumulative += dis[i] = randomValue() + 1;
     }
 
     // normalize
-    for (int i = 0; i < lon; i++) {
-        dis[i] /= acum;
+    for (int i = 0; i < len; i++) {
+        dis[i] /= cumulative;
     }
 }
 
 /*- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-	dis_inicDelta: genera una distribución aleatoria tipo izq-der con
+	dis_set_random_delta: genera una distribución aleatoria tipo izq-der con
 	desplazamiento máximo delta.
 */
-void dis_inicDelta(prob_t *dis, int lon, int desde, int delta) {
+void dis_set_random_delta(prob_t *dis, int len, int from, int delta) {
     int i, resto, numval;
 
     /* poner en ceros toda la distribución: */
-    for (i = 0; i < lon; i++)
+    for (i = 0; i < len; i++)
         dis[i] = 0.;
 
-    /* la asignación de valores no negativos comienza en el índice desde,
-     * desde donde hay lon-desde valores hasta el final de la distribución: */
+    /* la asignación de valores no negativos comienza en el índice from,
+     * desde donde hay len-from valores hasta el final de la distribución: */
 
-    resto = lon - desde;
+    resto = len - from;
 
     /* número válido para la asignación: */
     numval = min(resto, delta);
 
     /* generar distribución aleatoria para numval valores: */
-    dis_inicAle(dis + desde, numval);
+    dis_set_random(dis + from, numval);
 }
 
 // uniform distribution
-void dis_inicUni(prob_t *dis, int lon) {
-    prob_t val = (prob_t) 1. / lon;
+void dis_set_uniform(prob_t *dis, int len) {
+    prob_t val = (prob_t) 1. / len;
 
-    prob_t acum = 0.;
-    for (int i = 0; i < lon - 1; i++) {
-        acum += dis[i] = val;
+    prob_t cumulative = 0.;
+    for (int i = 0; i < len - 1; i++) {
+        cumulative += dis[i] = val;
     }
 
-    // el acumulado es por cuestiones de precisión.
-
-    dis[lon - 1] = 1 - acum;
+    dis[len - 1] = 1 - cumulative;
 }
 
 // returns an event as outcome of an experiment over the distribution
-int dis_event(prob_t *dis, int ldis) {
+int dis_event(prob_t *dis, int len) {
     prob_t r = randomValue();
 
     int e = 0;
-    while ((r > dis[e]) && (e < ldis)) {
+    while ((r > dis[e]) && (e < len)) {
         r -= dis[e++];
     }
 
-    return e < ldis ? e : --ldis;
+    return e < len ? e : --len;
 }