+++ src/ecoz2/ecoz2.c
@@ -29,7 +29,7 @@
 #endif
 
 const char *ecoz2_version() {
-    return "0.4.1";
+    return "0.4.2";
 }
 
 void ecoz2_set_random_seed(int seed) {